Completed searchFilter component of part 2.10
